Move accountsjs utils unit test closer to package
