+++ packages/core-accountsjs/jest.config.js
@@ -11,4 +11,7 @@ export default {
       },
     ],
   },
+  moduleNameMapper: {
+    '(.+)\\.js': '$1',
+  },
 };
