+++ packages/core-assortments/tests/assortments-index.test.ts
@@ -1,5 +1,5 @@
 import { Tree } from "@unchainedshop/types/common.js";
-import { buildFindSelector } from "../lib/module/configureAssortmentsModule.js";
+import { buildFindSelector } from "../src/module/configureAssortmentsModule.js";
 import { divideTreeByLevels, concatItemsByLevels, fillToSameLengthArray, fillUp } from "../lib/utils/tree-zipper/zipTreeByDeepness.js";
