+++ packages/core-countries/tests/countries-index.test.ts
@@ -1,4 +1,4 @@
-import {buildFindSelector} from "../lib/module/configureCountriesModule.js"
+import {buildFindSelector} from "../src/module/configureCountriesModule.js"
 describe('Country', () => {
   
   it('buildFindSelector should return correct filter object', async () => {
